Chore: Relocate Engine parameter type definitions to engine-types for reuse in other Rust contracts
