+++ engine-types/src/parameters/mod.rs
@@ -4,6 +4,8 @@ use crate::*;
 use borsh::maybestd::io;
 use borsh::{BorshDeserialize, BorshSerialize};
 
+pub mod engine;
+
 #[must_use]
 #[derive(Debug, BorshSerialize, BorshDeserialize)]
 pub enum PromiseArgs {
