+++ engine/src/errors.rs
@@ -1,3 +1,7 @@
+pub use aurora_engine_types::parameters::engine::errors::{
+    ERR_CALL_TOO_DEEP, ERR_OUT_OF_FUNDS, ERR_OUT_OF_GAS, ERR_OUT_OF_OFFSET, ERR_REVERT,
+};
+
 pub const ERR_NOT_A_JSON_TYPE: &[u8; 19] = b"ERR_NOT_A_JSON_TYPE";
 pub const ERR_JSON_MISSING_VALUE: &[u8; 22] = b"ERR_JSON_MISSING_VALUE";
 pub const ERR_FAILED_PARSE_U8: &[u8; 19] = b"ERR_FAILED_PARSE_U8";
@@ -32,11 +36,8 @@ pub const ERR_STACK_OVERFLOW: &[u8; 18] = b"ERR_STACK_OVERFLOW";
 pub const ERR_INVALID_JUMP: &[u8; 16] = b"ERR_INVALID_JUMP";
 pub const ERR_INVALID_RANGE: &[u8; 17] = b"ERR_INVALID_RANGE";
 pub const ERR_DESIGNATED_INVALID: &[u8; 22] = b"ERR_DESIGNATED_INVALID";
-pub const ERR_CALL_TOO_DEEP: &[u8; 17] = b"ERR_CALL_TOO_DEEP";
 pub const ERR_CREATE_COLLISION: &[u8; 20] = b"ERR_CREATE_COLLISION";
 pub const ERR_CREATE_CONTRACT_LIMIT: &[u8; 25] = b"ERR_CREATE_CONTRACT_LIMIT";
-pub const ERR_OUT_OF_OFFSET: &[u8; 17] = b"ERR_OUT_OF_OFFSET";
-pub const ERR_OUT_OF_GAS: &[u8; 14] = b"ERR_OUT_OF_GAS";
 pub const ERR_OUT_OF_FUND: &[u8; 15] = b"ERR_OUT_OF_FUND";
 pub const ERR_NOT_SUPPORTED: &[u8; 17] = b"ERR_NOT_SUPPORTED";
 pub const ERR_UNHANDLED_INTERRUPT: &[u8; 23] = b"ERR_UNHANDLED_INTERRUPT";
@@ -84,6 +85,3 @@ pub const ERR_FAILED_UNREGISTER_ACCOUNT_POSITIVE_BALANCE: &[u8; 46] =
     b"ERR_FAILED_UNREGISTER_ACCOUNT_POSITIVE_BALANCE";
 
 pub const ERR_ACCOUNTS_COUNTER_OVERFLOW: &str = "ERR_ACCOUNTS_COUNTER_OVERFLOW";
-
-pub const ERR_REVERT: &[u8; 10] = b"ERR_REVERT";
-pub const ERR_OUT_OF_FUNDS: &[u8; 16] = b"ERR_OUT_OF_FUNDS";
